refactor: move result to internals
