+++ ulauncher/internals/result.py
