+++ ulauncher/modes/shortcuts/ShortcutResult.py
@@ -2,9 +2,9 @@ from __future__ import annotations
 
 import re
 
-from ulauncher.api.result import Result
 from ulauncher.api.shared.action.OpenAction import OpenAction
 from ulauncher.internals.query import Query
+from ulauncher.internals.result import Result
 from ulauncher.modes.shortcuts.run_script import run_script
